+++ Source/Bake.Tests/IntegrationTests/BakeTests/NetCoreConsoleTests.cs
@@ -38,6 +38,21 @@ namespace Bake.Tests.IntegrationTests.BakeTests
         }
 
         [Test]
+        public async Task Minimal()
+        {
+            // Arrange
+            var testState = TestState.New("run");
+
+            // Act
+            var returnCode = await ExecuteAsync(testState);
+
+            // Assert
+            returnCode.Should().Be(0);
+            AssertSuccessfulArtifacts();
+            Releases.Should().HaveCount(0);
+        }
+
+        [Test]
         public async Task Run()
         {
             // Arrange
@@ -45,7 +60,7 @@ namespace Bake.Tests.IntegrationTests.BakeTests
                 "run",
                 "--print-plan=true",
                 "--convention=Release",
-                "--destination=nuget>http://localhost:5555/v3/index.json",
+                "--destination=release>github,nuget>http://localhost:5555/v3/index.json",
                 "--build-version", SemVer.Random.ToString())
                 .WithEnvironmentVariable("bake_credentials_nuget_localhost_apikey", "acd0b30512ac4fa39f62eb7a61fcf56c");
 
@@ -55,6 +70,7 @@ namespace Bake.Tests.IntegrationTests.BakeTests
             // Assert
             returnCode.Should().Be(0);
             AssertSuccessfulArtifacts();
+            Releases.Should().HaveCount(1);
         }
 
         [TestCase(LogEventLevel.Verbose)]
@@ -69,7 +85,7 @@ namespace Bake.Tests.IntegrationTests.BakeTests
             var planPath = Path.Combine(WorkingDirectory, "plan.bake");
             var returnCode = await ExecuteAsync(
                 "plan",
-                "--destination=\"nuget>github\"",
+                "--destination=release>github,nuget>github",
                 $"--log-level:{logLevel}",
                 "--build-version", SemVer.Random.ToString(),
                 "--plan-path", $"\"{planPath}\"");
@@ -95,6 +111,18 @@ namespace Bake.Tests.IntegrationTests.BakeTests
             AssertFileExists(
                 50L.MB(),
                 "bin", "Release", "publish", "win-x64", "NetCore.Console.exe");
+        }
+
+        [TestCase(ExitCodes.Core.NoCommand)]
+        public async Task ExitCodeTests(
+            int expectedExitCode,
+            params string[] args)
+        {
+            // Act
+            var returnCode = await ExecuteAsync(args);
+
+            // Assert
+            returnCode.Should().Be(expectedExitCode);
         }
 
         [TestCase("-h")]